Rename and clean up the routing function files
