+++ src/functions.jl
@@ -5,9 +5,8 @@ const DE = DifferentialEquations
 
 
 include("pseudo_witness_sets.jl")
-include("slice_gradient_hessian.jl")
-#include("routing_points.jl")
-include("slice_system.jl")
+include("gradient_cache.jl")
+include("routing_functions.jl")
 include("homotopy.jl")
 include("ode_solving.jl")
 include("graph.jl")
