+++ src/routing_functions.jl
